honeycomb_benches

Module shift

Description

Routine

The algorithm fetches all vertices that are not on the border of the map, fetch all identifiers of each respective vertices’ neighbors. Then, for all vertices:

compute the average between neighbors

overwrite current vertex value with computed average

Benchmark

This binary is meant to be use to evaluate scalability of geometry-only kernels. It is parallelized using rayon, and the number of thread used for execution can be controlled using taskset. By controlling this, and the grid size, we can evaluate both strong and weak scaling characteristics.

Functions

bench_shift